requirements.txt
streamlit
PyPDF2
python-docx
python-pptx
scikit-learn
nltk